+++ src/com/kaylerrenslow/armaDialogCreator/gui/main/controlPropertiesEditor/ControlPropertiesEditorPane.java
@@ -209,5 +209,4 @@ public class ControlPropertiesEditorPane extends StackPane {
 	}
 
 
-
 }