+++ src/com/kaylerrenslow/armaDialogCreator/gui/main/popup/newControl/NewControlDialog.java
@@ -46,7 +46,7 @@ public class NewControlDialog extends StageDialog<GridPane> {
 				if (group != controlType.getGroup()) {
 					continue;
 				}
-				groupMenu.getCbmbMenuItems().add(new CBMBMenuItem<>(controlType, new BorderedImageView(controlType.getIcon())));
+				groupMenu.getCBMBMenuItems().add(new CBMBMenuItem<>(controlType, new BorderedImageView(controlType.getIcon())));
 			}
 			if (groupMenu.getItems().size() > 0) {
 				menuButtonControlType.addGroup(groupMenu);